initial migration added

+++ GymBro.Domain/Entities/ExerciseEquipment.cs
@@ -12,7 +12,7 @@ namespace GymBro.Domain.Entities
     {
         public long Id { get; private init; }
         public long ExerciseId { get; set; }
-        public int EquipmentId { get; set; }
+        public long EquipmentId { get; set; }
 
         public Exercise? Exercise { get; set; }
         public Equipment? Equipment { get; set; }
